Attempted a "large" version of ddp-in-memory
Writing and reading the vector using the hard disk should already allow
us to reason about vectors that are larger than system memory. So all we
need to ensure is that actors never have the entire thing loaded at the
same time. This is the "trivial" way of achieving this: Simply split the
work into a few steps. This is probably nothing but a starting point
though - eventually we probably want this to happen without re-starting
actors, to say nothing about automatic scheduling...