Test linux-perf-stat for dot product implemented in Haskell, SSE and AVX.
This is actually important - we can verify that we can indeed
catch both SSE as well as AVX SIMD operations. We especially
get an idea of how much double-counting we can expect. Bad news
here is that it turns out to be substantial (~60%) for naive
SSE/AVX. Proper prefetching pushes this number below 10%.