linux-perf-stat supports PMU counters
This implements another way of naming an perf_event counter: Using
the entries in /sys. I first didn't want to go this route, as it
seemed redundant - but apparently there are a few counters that
only appear here (QPI, to be exact), so let's get them on board.

Also various clean-ups & documentation.