CUPTI metrics
This is basically the perf_events equivalent for CUDA - except that it
is a lot more fiddly to make work. We especially have to do a lot of
hooking in order to get them to activate at just the right time.
Furthermore, a lot of subtle things can go wrong, I tried to be careful
about generating warnings for all of them.