DNA support for CUDA measurements
Apart from the additional CUDA hints, the most user-visible change is that
we have now have a "--measure xxx" command line parameter, which can be
used to request certain measurements. Note that this generally goes at
the expense of performance! Currently the CUDA profiling module supports
"fp-inst", "float-ops" and "double-ops".

Internally, we are now activating and deactivating modules as appropriate.
Also, the duplicated hint in both START and END message was fixed.