Merge "kernel" and "profile", remove "liftIO"
Furthermore, running in bound threads is now the default for kernels,
use "unboundKernel" for the old behavior.

This enforces a strict separation between IO and DNA. Upsides:

 * Hopefully more robust (CUDA?)
 * Easier to do meaningful profiling
 * All IO is automatically profiled in some fashion

The heavy disadvantage is that now you even have to profile "trivial" IO,
such as memory management. It might well be that we'll eventually
reintroduce liftIO in a way that automatically suppresses profile
generation for trivial costs.