Yet more visualisation improvements
Now stacking bars, which is a lot more compact and therefore useful.

On the other end of the spectrum, I'm now showing the internally-used
reference times. Makes the overview rather busy, but there's been so
much confusion about them that we probably won't get around them. Splitting
up the table is probably the next step.