Make perf_events normalise counter values
Not completely sure whether it would make more sense to just write
"running" and "enabled" into the eventlog and have the analysis tool
figure out the rest. However the only thing this could ever be used for
would be normalisation - or possibly warning about the loss of accuracy
due to normalisation happening. So doing it here seems okay.

Also refactoring, prepararing for more counters.