Memory performance counters
The counter of choice is the "offcore response" last-level-cache counter.
This follows the choice that the Linux kernel makes in newer versions when
we ask for read misses on LLC. Our kernels are old, however, so we need
to hard-code the values instead.

Introduces a new performance hint (memHint), which can be used to
document expected memory transfer volume. Keep in mind that having
"floatHint" and "memHint" on the same kernel means that the kernel will
start multiplexing performance counters, at a loss of accuracy.