First draft of top-level program
It is not particularly pretty, especially note that "Vector" is just
wrapping pointers, which we have to free ourselves. A lot of functions
will consume their input. This is very un-Haskell-y, but we *do* care
about allocations and de-allocations. A lot.

On the other hand, we have a minimal kernel abstraction now. Tried to
keep as close as possible to current kernel function, but this will
probabaly still need some major revisions sooner or later.