Configuration
We use Aeson (= JSON) for the configuration files. This might be
overkill in terms of dependencies, but it gives us an extensible
starting point. Especially because the whole polarisation debate
doesn't seem to be settled yet.

Also moved all configuration to its own module. No need to pollute
imaging.hs with the parsing details.