Cabalified "oskar" library
Idea is to have all the oskar code in the repository so we don't need
to include external code to build. If Peter really wants a one-button
compilation, that's pretty much required.

This also allows us to deal with all the mess of linking a C library
in a central fashion. This should keep C (and especially C++) dependencies
out of the top-level program.

Finally, I tried to streamline the interface a bit, removing baseline
maps and passing through the metrics. Hopefully I didn't break too much,
I am currently lacking Oskar data to test this.