Eliminate lambda parameter
If I read crocodile right, it can be derived from the grid size and
theta. Multiplying out the dependency actually simplifies uv scaling
calculations, so I'll simplify this for the moment.  If we find out later
that this doesn't make sense, we'll just revert, but for the moment it
simplifies finding good configurations a great deal.