Use Kern for carrying MonadLog data
Before, most of DNA.Logging was actually running in IO and therefore just
had access to the trivial "default" MonadLog implementation. This
basically meant that it was impossible to access command line parameters.

Root of the problem was that execKernel had no way to re-introduce the
logging parameters into the IO passed to async/asyncBound. So what I do
here is to introduce the required data into Kern and run logProfile in
that monad.

While we're at it, this also propagates this change further to replace
"IO" by "Kern" pretty much on the whole DNA interface. After all, this
make it consistent and now actually has some marginal usefulness...!