Documentation
I added a dummy cabal file just for the dna programs. It is purely useful
for running

  cabal --sandbox-config-file=../cabal.sandbox.config haddock --executables

To generate documentation for the executables. This works around a bug
in cabal where you cannot use --executables for something that depends on
the "local" library apparently (see github.com/haskell/cabal/issues/1919).