Mock-Up 4: Rules
The idea is that instead of manually binding kernels in accordance with
the data flow graph, we formulate rules that match the data flow graph
and are therefore guaranteed to fit in. More guarantees means less
restrictions when programming, so that's good.

Rules aren't quite type-safe yet, and there is a *lot* of potential for
subtle errors here if the magic goes bad somewhere. But on the other hand,
this makes the strategy basically collapse into just a few lines, which
are all about the few cases where the automatic matching doesn't quite
do the right thing (= don't memoise the result of "create grid").

At least partly inspired by Kyrill calling my approach "graph rewriting".