Mockup 8: Eliminate flow input parameter again
After thinking a bit about it, maybe we don't actually need this so much.
Instead, we now declare explicitly what the "input" for every kernel is,
and we just check whether it is some sort of dependency in the abstract
data flow. Strictly speaking this reduces static guarantees a bit, this
is very hard to realise.

On the other hand, this has some significant advantages even if we
ignore the shorter notation:

1. It is now more straightforward to "rebind" the value of a kernel. You
   simply declare the kernel to produce X from X, and specify the same
   flow for input as well as output.

2. Now we can actually have kernels implement non-trivial data flows at
   once. One example is PSF, which is most cleanly described as gridding
   over modified visibilities, but realistically we might want to simply
   "fuse" this into a modified gridding kernel. This is now possible to
   express.