Kernel input list currying
Saving the input flows alongside the kernel that uses it works remarkably
well - after some heaviy lifting on the type level, "Kernel" becomes
something we can simply pass to "prepareKernel". Furthermore, on the
front end this means we get a consistent syntax for input flows, no matter
whether we are constructing

1. a new primitive flow,
2. a compound flow (flow functions) or
3. a kernel

So now we can actually write

  bind1D dh (gcf tag vis) (gcfKernel gpar tag vis)

Which is probably as obvious as we can make it.

Downside: The type magic in Builder is a mess, and probably can be
simplified. The names are even worse. Plus this took me *way* too long
to figure out. I'll have to start working on something more practical.