Make "tag" flow strategy-only
Actually realised that there is no reason to have it in the abstract data
flow: As long as we make sure that it is a parent of the "vis" data flow,
it is an ancestor of every node in the abstract data flow, and therefore
a valid kernel data dependency.

Also split out the major loop iteration.