First (rather hacky) attempt at domains
This is a rough first stab at implementing range domains, which is what
we need at minimum in order to make distributed dot product work. The most
notable design decision here is that we pull the domain into the *data
representation*. This makes a lot of sense - a distributed vector is not
the same thing as a non-distributed vector. Kernels just inherit this
property from their output data representation. Not quite sure yet how to
handle kernels that are expected to produce a distributed data structure
but aren't called in a distributed fashion (automatic sequential
distribution?).

There's a lot still left to resolve here - it's awfully specialised to
one-dimensional ranges, and dotproduct especially gives the wrong result
due to the domain offset not getting communicated to Halide. We'll have
to see about that.