Pass domain offset to Halide functions
This is basically the last piece of the puzzle for distributed dot
product. Unfortunately, it again involves some type trickery and doesn't
quite solve the whole problem, but well.

A tricky aspect here is that now the Halide function type depends on the
data representation of the return type - because if the data representation
allows for splitting regions we need to communicate which one we want
generated. This makes the FFI definitions a bit less obvious to get
right.