Scatter gridding kernel
The bindings are wrong at this point, as the grid shouldn't be a parameter,
but instead simply be passed in as existing data in the output buffer.
We will have to figure out exactly how to formulate that as a dat
representation.

But in theory this should be most of the groundwork we need to get a first
gridder-like pipeline going.