Have IO interpreter check buffer sizes
Data representations should normally know what the "right" buffer size
is supposed to be, so we can automatically check it.

This actually showed that I was mis-calculating the visibility buffer
size, leading to the segfault before. Now some sort of gridding runs
through, but we still need to sort out FFT.