Added Halide FFT
The FFT from Halide's repository only uses single-precision floating
point numbers, so I had to adapt it a bit. Still produces a lot of warnings
due to missing casts and takes *ages* to generate the kernel, but it
seems to run through okay.

The result image doesn't look quite like we'd expect, but that would have
been a small miracle anyway.