Push image resolution to 8K x 8K
Also change the grid theta so we have enough detail to put into it.
This moves two more dots from the test data into the picture.

The trickiest part of this is to make sure that the Halide FFT compile
time doesn't completely explode. Turns out they were a bit gratious with
the unrolling (especially transpose), and a bit of adjustment makes things
much nicer overall.

While I'm at it, I also merged the "unzipped" loop with our "shifted"
loop. They both simply rearrange the data, so this should be a strict win.
Not completely convinced that LLVM is smart enough to make everything
efficient, but to my untrained eye it looks pretty good.