Move special domain handling into kernels
Makes the KernelCode quite a bit more powerful (read: complicated), but
in this case that's a clear win. We move a lot of weirdness out of the
IO scheduler. The new "mappingKernel" and "mergingKernel" capture this in
a clearer way, while leaving options open for kernels to do more
interesting things with data that comes or is expected in a region-split
form.