Added region data representation
This is an easy data representation that is split according to the
regions of a domain, without actually caring about it in terms of data
structure or function. Generally, this is because we have produced it
from or for data that is structured in a certain way, and want to improve
locality. The driving example here is to have GCF per bin, but it could
also be the sum per vector slice in a tree reduction.