Folding kernels
Just like with mapping kernels, we run the given code multiple times in
order to get the final result. However this time we run over multiple
parameter regions while still producing only one result (= a reduction),
so we need a kernel that can be applied to the same data as many times
as we have regions in the parameters.

First wanted to do this as a kernel combinator, but that doesn't really
work out too well. This does okay, and especially fits well with the
Halide "write"-type kernels.