Use margin representation for uv-tiling
At this point, all we have to do is use the "margin" data representation
for splitting up the uv-grid.

The tricky part is merging the uv-tiles together again,
especially given that MarginRepr isn't automatically mergeable as explained
in the last commit. But this isn't particularly hard to get around, as
we can actually write a Halide kernel that updates the grid from tiles
using the exact same machinery that w-tiling uses. This is probably
slightly over-engineering things (and we are cheating a bit with
"createGrid"), but I couldn't resist.

Especially note that the Halide kernel needs to be careful with setting
bounds, as including margins our field is actually *larger* than the
target field. In our case, this is actually a very good idea, as it allows
us to have GCFs partly overlapping the full grid's borders.