regionKernel modifier for calling kernels per-region
Can't shake the feeling that we'll want kernel modifiers/combinators
sooner or later. The example in imaging should show why: We want to
introduce a domain for different data sets, but once we have read the
data in the kernels actually don't care about that anymore. So it is
most natural to just "impose" the region on top of the kernel after the
fact.

This is quite an experiment, and the glue code in regionKernel turns out
to be quite messy. Maybe