Region-dependent bins
We are going to need this for defining bins that are filtered by a
secondary domain (say, uv-tile) in addition to the bin itself. This works
by having a "filter" that removes non-matching bins from the bin region
in presence of dependent regions. To allow region matching, we adjust
equality and comparison checks to ignore such changes to the bins.