Add uv-binning to gridder
At this point, we have a fairly confusing number of domains in play, and
a bunch of them apply to multiple data representations at once:

 1. The visibility domain (for raw visibilities)
 2. The full U range domain
 3. The full V range domain
 4. The split U range domain
 5. The split V range domain
 6. The full W bin domain
 7. The split W bin domain

The binner splits visibility data into the cartesian product of all split
domains (4,5 and 7), resulting in a fairly large number of individual
regions. The gridKernel will use GCFs (split by 7) to produce an UVG
grid tile (split by 4 and 5) which the detiling kernel integrates to
produce a full grid (2 and 3).

Getting all buffers to appear for the correct kernels was quite tricky to
get right... Fortunately the result is perfectly okay, so it seemingly
works out.