Eliminate bins with zero size
Assuming that this is correct is a slight leap, even though it is clearly
true for our examples: Empty visibility bins in either direction can be
ignored.

However, we need to make sure that the "Map keys" method used by the Eq
and Ord instances still works, so we should not eliminate individual
bins within bin regions, but always the complete region (box).