Merge SplitStep into DomainStep
At this point, this is a rather obvious improvement. I kept expecting
for "split" to have some sort of scope-like behaviour, but it hasn't
really materialised. And given that sooner or later we'll probably get
splits that are informed by kernel data (scheduling?) we *really* want
to be able to use the same infrastructure.