Propagate split change to DSL
Makes the whole program somewhat flat and imperative-looking, but well,
that's what you get with monads. At least we are in good company -
Halide kernels also quickly become collections for directives that are not
straightforward to untangle.

At least it now visually highlights the bits where we *do* distribute.
I threw in an extra "distribute" on the GCFs even though it doesn't
make any difference for execution due to the underlying "mappingKernel".