OpenMP JIT benchmark
This took more fiddling than anticipated - the Halide JIT interface
is apparently not thread-safe, so we need to bypass it and call the
kernel directly. This is complicated by the fact that for JIT Halide
"infers" the argument order and adds an extra "JITUserContext" parameter.
All this has to match up exactly or the kernel will just crash.

Furthermore, using clock() is unreliable and not very exact. Using
clock_gettime is to my knowledge the "proper" way for Linux, so let's use
that instead. Probably not portable, apologies.