Chunks, baselines and different GCF sizes for JIT benchmarks
Yet more variables to tweak - measurements suggest that for certain
GCF sizes it is a good idea to pull out the Y loop (as suggested by
Romein).

Furthermore, cost depends quite a bit on the baselines processed (with
central and out-of-bounds ones being cheap), so we need to allow
OpenMP a bit of load balancing in order to be efficient.