Pass region start to Halide for RegionRepr
This is yet another "hack" that makes sense if you squint the right way:
Region representations are only one instance per region, so it has to
be extent 1 to halide. However, we can still use the "min" field to
tell Halide *which* region we are currently looking at.

At least for range regions. If we try to do this for Bins, we get the
problem that GCFs might now end up with a grand total of 7 dimensions...