Create bin domain from buffer containing sizes, automatically size W domain
This trades a nice interface at the top for slightly more magic at the
low level. The binSizer Haskell kernel now especially needs to produce
an array containing both Int64 as well as Double, which isn't particularly
nice.

On the other hand, now the gridSizer has free reign to set the bins in
whatever way the data dictates. I especially took the opportunity to
clean up the binning code a bit more - especially so it doesn't count
visibilities that fall outside of our grid.