Stack build file
Experimental - this might allow us to remove all the cabal sandbox stuff.
With stack, this all boils down to just "stack build".

The main price we pay is that we are more restricted in the packages we
are allowed to depend on. See CUPTI - now the tests might not work any
more due to a mismatch in the accelerate package. Might have to look into
that later.