Add scheduling
Not fully functional yet, but this is the bulk of the infrastructure
we'll need for scheduling. Idea here is that we use binning domains to
do the distribution work for us: makeScheduleDomain creates a domain
with one bin per data set, and a matching Flow that carries the indices
of the data sets as data.

From there, we can split the bin domain using a simple "split", and
use a reasonably straight-forward slicing kernel to restrict the Flow
to exactly the data set indices we are supposed to work on. Note that
this only works because the input configuration is global, so we can
consider indices to actually identify data sets - otherwise we would have
to fill the index Flow with file names or something to that effect.

Worst part of this is that data representation types are shared between
the test programs, which basically means that even programs running
over only one data set need to get most of the scheduling "plumbing".