Fix stepsKernDeps for DistributeStep
Dependencies satisfied inside a DistributeStep were ignored. Normally this
doesn't matter too much, as we only use it for GC, which would at maximum
keep data around for too long. However, in execDistributeStep we actually
take a set difference, and this means that it matters.

This fixes a problem with buffers getting GC'd too early with deep
distribution loops.