Complete major loop iteration
This adds the missing pieces to complete a full major loop iteration.
This formulation is actually closer to the SKA top-level data flow
than what we had in "imaging":

 1. We re-read visibilities every single iteration. This makes visibilities
    completely internal to the inner continuum gridding loops.

 2. The first major loop actually de-grids an empty model, which means
    the visibilities just get returned unchanged.

While this might seem puzzling, this actually makes a lot of sense: For
the full-scale SKA, we will have terabytes worth of visibilities, and
just loading them locally on-demand will absolutely be the right approach.
Plus instead of the empty start model we would actually have a global
sky model that we can use to "seed" the loop.