Add facetting to continuum
The trickiest problem here is a fairly technical issue when scattering
the image to use for degridding. Normally we'd just declare it as
coming from the "global" domain to the kernel - however when using
"regionKernel" we add the domain in question on *all* kernel inputs
including the image. On the other hand, if we add "RegionRepr" manually
we'd actually end up with a 7-dimensional array on the Halide side.

So instead, we have a simple glue kernel that simply "lifts" the image
into the appropriate region domains. In order to do this automatically
we need "rule" from the flow builder.