I/O hints
The data representation data actually allows us to predict transfer sizes
very accurately. In fact, for Halide-type data representations we know
the data object sizes down to the byte once we have the appropriate
region information.

Making this work still requires quite a bit of plumbing however. We need
the ability to query a kernel's data representation ahead of time, and
must be able to identify the right domain information to pass to it.

The back-channel to the parent actor is especially tricky here - even
though the parent theoretically knows the exact domain information that
was sent to the child (localDomainMap), DNA doesn't actually allow us to
identify the client that sent data back. We use the easy way out here
and simply have clients send their domain map back - even though that's
actually decidedly redundant.