Call actor code directly for sequential composition
This is essentially falling back to how the pure IO interpreter was doing
things - simply pass a restricted map of buffers to the client actor
without bothering with copies. While this is a lot more efficient, it
is also a lot tricker to get right - we must make sure that the client
doesn't end up freeing buffers that the parent (or another child being
executed sequentially) might still need.

To keep things simple, this implements the most simple variant: Declare
that children are not allowed to free their inputs. This keeps buffers
alive for too long, but other solutions would need trickier dynamic
reasoning, which is a bit against the spirit here.

Note that to make this work we also need to implement duplication of
written kernel inputs. This has been on the ToDo list for quite some time
already, and actually turns out to be relatively easy to implement.