w-depdendent GCFs
Solution is fairly straightforward - "just" yet another use for bins.
This time, we use two bin domains to set the size of the GCF data
representation depending on what w-region we are currently in. Given the
whole dependent domain machinery this is really easy to get right.

What this gives us is the basic cost shape of how gridding would work. The
GCFs still aren't realistic, we really ought to conjugate them for negative
w, and with the way margin currently works, we are probably gridding
a few visibilities on facet borders twice. But it's close enough to get
us started.