Change margin handling to account for w-dependent GCF sizes
Before we were relying on the grid margin always matching the GCF size
to make sure that every visibility was only gridded once. Now that GCFs
might be smaller, we have to skip some GCFs that would actually fit
into the margins to prevent duplication.

Not particularly complicated, but we need to pass an extra parameter to
all de/gridding kernels.