Conjugate w-kernel for negative w
Cheap way to get this roughly right. Note that for small w this isn't
quite right, but I think (?) there's no significant imaginary portion
in that GCF anyway. It doesn't seem to make a difference for the output,
but we should at least acknowledge that this is the right thing to do.