Fix residual output
Our residual output has actually been an empty image for quite some time.
Reason was simply parameter order - cleanResidual received the image
as PSF and an (empty) model as image. Hardly surprising that the output
residual image was empty as well.

This can actually easily be fixed by just swapping the rule parameter
order - and the type checker would have caught it if we had not given the
model the Image tag. For the moment, it's a bit subtle, so I added a
comment.